rename helloworld-jersey to onboarding-enabler-sample-app fix documentation
