+++ gradle/coverage.gradle
@@ -8,7 +8,8 @@ ext.javaProjectsWithUnitTests = [
     'gateway-service',
     'onboarding-enabler-java',
     'onboarding-enabler-spring',
-    'helloworld-jersey',
+    'onboarding-enabler-java-sample-app',
+    'integration-enabler-spring-v1-sample-app',
     'security-module',
     'security-service-client-spring',
     'apiml-security-common'
