--- helloworld-jersey/src/main/resources/openapi-configuration.yaml
@@ -1,11 +1,0 @@
-resourcePackages:
-    - org.zowe.apiml.hwsjersey
-prettyPrint: true
-cacheTTL: 0
-openAPI:
-    info:
-        version: '1.0.0'
-        title: Hello World Jersey
-        description: 'REST API for a Jersey Application'
-    servers:
-        - url: https://localhost:10016/hellojersey
