+++ onboarding-enabler-java-sample-app/src/main/java/org/zowe/apiml/hwsjersey/greeting/Greeting.java
