+++ onboarding-enabler-java-sample-app/src/main/resources/service-configuration.yml
@@ -1,7 +1,7 @@
-serviceId: hellojersey
-title: Hello Jersey REST API
-description: Example for exposing a Jersey REST API
-baseUrl: https://localhost:10016/hellojersey
+serviceId: enablerjavasampleapp
+title: Onboarding Enabler Java Sample App
+description: Example for exposing a Jersey REST API using Onboarding Enabler Java
+baseUrl: https://localhost:10016/enablerJavaSampleApp
 serviceIpAddress: 127.0.0.1
 
 homePageRelativeUrl:
@@ -13,14 +13,14 @@ discoveryServiceUrls:
 
 routes:
     - gatewayUrl: api/v1
-      serviceUrl: /hellojersey/api/v1
+      serviceUrl: /enablerJavaSampleApp/api/v1
     - gatewayUrl: ui/v1
       serviceUrl: /
 
 apiInfo:
-    - apiId: org.zowe.apiml.hellojersey
+    - apiId: org.zowe.apiml.enablerJavaSampleApp
       gatewayUrl: api/v1
-      swaggerUrl: https://localhost:10016/hellojersey/openapi.json
+      swaggerUrl: https://localhost:10016/enablerJavaSampleApp/openapi.json
 
 catalog:
     tile:
