+++ onboarding-enabler-java-sample-app/src/main/webapp/index.jsp
