+++ onboarding-enabler-java-sample-app/src/test/resources/service-config.yml
