+++ settings.gradle
@@ -1,6 +1,5 @@
 rootProject.name = 'api-layer'
 
-include 'helloworld-jersey'
 include 'discovery-service'
 include 'apiml-utility'
 include 'apiml-common'
@@ -9,6 +8,7 @@ include 'gateway-service'
 include 'common-service-core'
 include 'discoverable-client'
 include 'integration-tests'
+include 'onboarding-enabler-java-sample-app'
 include 'integration-enabler-spring-v1-sample-app'
 include 'zowe-install'
 include 'api-catalog-services'
